+++ LIBFT/ft_printf/ft_uitoa_base.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 #include "../libft.h"
 
-int	ft_uitoaprint_base(unsigned int value, int base, int fd)
+int	ft_uitoaprint_base(uintmax_t value, int base, int fd)
 {
 	char	buf[50];
 	char    *ptr;
@@ -25,7 +25,7 @@ int	ft_uitoaprint_base(unsigned int value, int base, int fd)
 	return (ft_strlen(ptr));
 }
 
-int	ft_uitoalen_base(unsigned int value, int base)
+int	ft_uitoalen_base(uintmax_t value, int base)
 {
 	char	buf[50];
 	char    *ptr;
